requirements.txt
numpy==1.23.5
pandas==1.5.3
plotly==5.11.0
matplotlib==3.7.1
pytest==7.2.1
flake8==6.0.0
requests==2.28.1
scikit_learn==1.1.3
streamlit==1.18.1
plotly==5.11.0
s3fs==2023.1.0
PyYAML==5.4.1
bertopic==0.14.1
tqdm==4.64.1
Jinja2==3.1.2
xgboost==1.7.4
